pyproject.toml
[tool.poetry]
name = "ChatICommentsVk"
version = "0.1.0"
description = ""
authors = ["Rasul Taimaskhanov <81189712+taimaskhanov11@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.6.0"
aiovk = "^4.0.0"
aiogram = "^2.19"
pydantic = "^1.9.0"
PyYAML = "^6.0"
aioredis = {extras = ["hiredis"], version = "^2.0.1"}


[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "*"
isort = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
